Move utility traits to traits folder
